Revert "Creating password form modal and improving profile form (#253) (#14)"
This reverts commit f67a1f54fa8bc92e08dd8939f53a370ddd91dc9b.
