+++ src/app/profile/components/ProfileForm.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import axios from 'axios';
 import { useSession } from 'next-auth/react';
 import { useEffect, useState } from 'react';
@@ -14,14 +12,10 @@ import { PencilIcon } from '@/icons';
 import { useFormSubmission, useModal, useToast, useValidatedFormData } from '@/hooks';
 import { requiredFieldRule } from '@/shared/utils';
 
-import PasswordFormModal from './PasswordFormModal';
-
 export default function ProfileForm() {
 	const [fetchLoading, setFetchLoading] = useState(false);
 	const [error, setError] = useState<string | null>(null);
-	const [isEditing, setIsEditing] = useState(false);
 
-	const passwordFormModal = useModal();
 	const deleteAccountModal = useModal();
 	const deletePhotoModal = useModal();
 	const uploadModal = useModal();
@@ -43,22 +37,22 @@ export default function ProfileForm() {
 	});
 
 	// Fetch data
-	const fetchProfileData = async () => {
-		setFetchLoading(true);
-		try {
-			const response = await axios.get('/api/profile');
-			setValues(response.data);
-		} catch (error) {
-			console.error('Error loading profile data:', error);
-			setError('Failed to load profile data! Please try again later.');
-		} finally {
-			setFetchLoading(false);
-		}
-	};
-
 	useEffect(() => {
+		const fetchProfileData = async () => {
+			setFetchLoading(true);
+			try {
+				const response = await axios.get('/api/profile');
+				setValues(response.data);
+			} catch (error) {
+				console.error('Error loading profile data:', error);
+				setError('Failed to load profile data! Please try again later.');
+			} finally {
+				setFetchLoading(false);
+			}
+		};
+
 		fetchProfileData();
-	}, []);
+	}, [setValues]);
 
 	// Submit data
 	const { loading, handleSubmit, toast } = useFormSubmission({
@@ -84,7 +78,6 @@ export default function ProfileForm() {
 							message: 'Profile updated successfully!',
 							variant: 'success',
 						});
-						setIsEditing(false);
 					}
 				} catch (error: unknown) {
 					// Narrowing down the type of `error`
@@ -126,15 +119,6 @@ export default function ProfileForm() {
 			}
 		},
 	});
-
-	const handleEditProfileInfo = () => {
-		setIsEditing(true);
-	};
-
-	const handleCancelEditing = () => {
-		setIsEditing(false);
-		fetchProfileData();
-	};
 
 	const handleDeleteAccount = () => {
 		console.log('Account deleted!');
@@ -185,62 +169,6 @@ export default function ProfileForm() {
 				autoComplete='off'>
 				<Grid
 					container
-					columnSpacing={{ xs: 1, sm: 2, md: 3 }}
-					alignItems='center'>
-					{/* Profile title */}
-					<Grid size={6}>
-						<Typography
-							variant='h1'
-							color='textPrimary'>
-							Profile
-						</Typography>
-					</Grid>
-					<Grid size={6}>
-						{/* Edit, Save and Cancel buttons */}
-						<Box
-							display='flex'
-							mb={5}
-							mt={3}
-							justifyContent='flex-end'>
-							{isEditing ? (
-								<Box
-									display={'flex'}
-									gap={4}>
-									<Button
-										variant='outlined'
-										color='secondary'
-										size='small'
-										onClick={handleCancelEditing}>
-										Cancel
-									</Button>
-									<LoadingButton
-										loading={loading}
-										size='small'
-										buttonText='Save'
-										loadingText='Saving...'
-										fullWidth={false}
-									/>
-								</Box>
-							) : (
-								<Button
-									variant='contained'
-									size='small'
-									onClick={handleEditProfileInfo}>
-									Edit
-								</Button>
-							)}
-						</Box>
-					</Grid>
-
-					{/* Divider */}
-					<Grid size={12}>
-						<Divider />
-					</Grid>
-				</Grid>
-
-				<Grid
-					container
-					my={10}
 					rowSpacing={14}
 					columnSpacing={{ xs: 1, sm: 2, md: 3 }}
 					alignItems='center'>
@@ -254,7 +182,6 @@ export default function ProfileForm() {
 							value={values.firstName}
 							onChange={handleChange}
 							errorMessage={getError('firstName')}
-							disabled={!isEditing}
 						/>
 					</Grid>
 
@@ -268,7 +195,6 @@ export default function ProfileForm() {
 							value={values.lastName}
 							onChange={handleChange}
 							errorMessage={getError('lastName')}
-							disabled={!isEditing}
 						/>
 					</Grid>
 
@@ -289,10 +215,8 @@ export default function ProfileForm() {
 						/>
 					</Grid>
 
-					{/* TODO: Avatar upload UI temporarily removed */}
-
 					{/* Photo */}
-					{/* <Grid size={6}>
+					<Grid size={6}>
 						<Typography variant='h4'>Your photo</Typography>
 						<Typography variant='subtitle1'>
 							This photo will be displayed on your profile page.
@@ -320,124 +244,97 @@ export default function ProfileForm() {
 									sx={{ width: 64, height: 64, mr: 7 }}
 								/>
 
-								{isEditing && (
-									<Box
-										className='avatar-edit-icon'
-										sx={{
-											position: 'absolute',
-											top: 0,
-											left: 0,
-											width: '100%',
-											height: '100%',
-											display: 'flex',
-											alignItems: 'center',
-											justifyContent: 'center',
-											bgcolor: 'rgba(0, 0, 0, 0.15)',
-											opacity: 0,
-											transition: 'opacity 0.3s',
-											cursor: 'pointer',
-										}}
-										onClick={uploadModal.openModal}>
-										<PencilIcon
-											width={20}
-											height={20}
-											color='white'
-										/>
-									</Box>
-								)}
+								<Box
+									className='avatar-edit-icon'
+									sx={{
+										position: 'absolute',
+										top: 0,
+										left: 0,
+										width: '100%',
+										height: '100%',
+										display: 'flex',
+										alignItems: 'center',
+										justifyContent: 'center',
+										bgcolor: 'rgba(0, 0, 0, 0.15)',
+										opacity: 0,
+										transition: 'opacity 0.3s',
+										cursor: 'pointer',
+									}}
+									onClick={uploadModal.openModal}>
+									<PencilIcon
+										width={20}
+										height={20}
+										color='white'
+									/>
+								</Box>
 							</Box>
 							<Link
 								href='#'
 								underline='hover'
 								pl={10}
 								color='text.secondary'
-								onClick={deletePhotoModal.openModal}
-								sx={isEditing ? {} : { pointerEvents: 'none', opacity: 0.5 }}>
+								onClick={deletePhotoModal.openModal}>
 								Delete
 							</Link>
 							<Link
 								href='#'
 								underline='hover'
 								px={8}
-								onClick={uploadModal.openModal}
-								sx={isEditing ? {} : { pointerEvents: 'none', opacity: 0.5 }}>
+								onClick={uploadModal.openModal}>
 								Update
 							</Link>
 						</Box>
-					</Grid> */}
-
-					{/* Password */}
-					<Grid
-						size={6}
-						mt={10}>
-						<Typography variant='h4'>Password</Typography>
-					</Grid>
-					<Grid
-						size={6}
-						mt={10}>
-						<Button
-							variant='contained'
-							// fullWidth
-							size='medium'
-							onClick={passwordFormModal.openModal}
-							disabled={!isEditing}>
-							Change password
-						</Button>
 					</Grid>
 				</Grid>
 
-				<Divider sx={{ mb: 10, mt: 20 }} />
+				{/* Save Button */}
+				<Box
+					display='flex'
+					justifyContent='flex-end'
+					mt={40}>
+					<LoadingButton
+						loading={loading}
+						buttonText='Save'
+						loadingText='Saving...'
+						fullWidth={false}
+					/>
+				</Box>
+
+				<Divider sx={{ mb: 7, mt: 14 }} />
 
 				{/* Delete Account Section */}
 				<Box
 					display='flex'
 					flexDirection='column'
 					mb={4}
-					rowGap={2}>
+					rowGap={6}>
 					<Typography variant='h4'>Delete account</Typography>
-					<Typography variant='subtitle1'>
-						Note that deleting your account will remove all data from our system. This is permanent
-						and non-recoverable.
-					</Typography>
 					<Typography
 						variant='subtitle1'
-						my={5}>
-						To Delete your account, please email us at{' '}
-						<Link
-							href='mailto:dev.datahall@gmail.com'
-							underline='hover'
-							color='primary'>
-							dev.datahall@gmail.com
-						</Link>
-						.
+						mb={2}>
+						Note that deleting your account will remove all data from our system. This is permanent
+						and non-recoverable.
 					</Typography>
 
 					{/* Delete Account Button */}
 
-					{/* <Tooltip
+					<Tooltip
 						title='Account deletion is disabled in Development'
 						placement='bottom-start'>
 						<Box width='9rem'>
 							<Button
 								variant='contained'
 								color='error'
-								size='medium'
 								onClick={deleteAccountModal.openModal}
 								disabled={true}>
 								Delete account
 							</Button>
 						</Box>
-					</Tooltip> */}
+					</Tooltip>
 				</Box>
 			</Box>
 
-			{/* Password form modal */}
-			<PasswordFormModal
-				open={passwordFormModal.isOpen}
-				toggleModal={passwordFormModal.closeModal}
-			/>
-
-			{/* Delete photo modal */}
+			{/* Delete Photo Modal */}
 			<ModalWrapper
 				variant='delete'
 				title='Really delete this photo?'
@@ -448,7 +345,7 @@ export default function ProfileForm() {
 				toggleModal={deletePhotoModal.closeModal}
 			/>
 
-			{/* Upload photo modal */}
+			{/* Upload Photo Modal */}
 			<ModalWrapper
 				variant='upload'
 				title='Upload profile image'
@@ -459,8 +356,7 @@ export default function ProfileForm() {
 				fileFormats='JPG, PNG'
 				toggleModal={uploadModal.closeModal}
 			/>
-
-			{/* Delete account modal */}
+			{/* Delete Account Modal */}
 			<ModalWrapper
 				variant='delete'
 				title='Really delete this account?'
